requirements.txt
torch>=2.4
torchvision>=0.19
opencv-python>=4.10
numpy>=1.26
pillow>=10.4
pyyaml>=6.0
tqdm>=4.66
einops>=0.8
playwright>=1.48
pandas>=2.2
scikit-learn>=1.5
matplotlib>=3.9
screeninfo>=0.8
